Completed unit tests and checks
