+++ tests/test_scraamp.py
@@ -1,116 +1,9 @@
 import numpy as np
 import numpy.testing as npt
-from stumpy import scraamp, core, aamp, config
+from stumpy import scraamp, aamp, config
 from stumpy.scraamp import prescraamp
 import pytest
 import naive
-
-
-def naive_prescraamp(T_A, m, T_B, s, exclusion_zone=None, p=2.0):
-    distance_matrix = naive.aamp_distance_matrix(T_A, T_B, m, p)
-
-    n_A = T_A.shape[0]
-    l = n_A - m + 1
-
-    P = np.empty(l)
-    I = np.empty(l, dtype=np.int64)
-    P[:] = np.inf
-    I[:] = -1
-
-    for i in np.random.permutation(range(0, l, s)):
-        distance_profile = distance_matrix[i]
-        if exclusion_zone is not None:
-            naive.apply_exclusion_zone(distance_profile, i, exclusion_zone, np.inf)
-        I[i] = np.argmin(distance_profile)
-        P[i] = distance_profile[I[i]]
-        if P[i] == np.inf:  # pragma: no cover
-            I[i] = -1
-        else:
-            j = I[i]
-            for k in range(1, min(s, l - max(i, j))):
-                d = distance_matrix[i + k, j + k]
-                if d < P[i + k]:
-                    P[i + k] = d
-                    I[i + k] = j + k
-                if d < P[j + k]:
-                    P[j + k] = d
-                    I[j + k] = i + k
-
-            for k in range(1, min(s, i + 1, j + 1)):
-                d = distance_matrix[i - k, j - k]
-                if d < P[i - k]:
-                    P[i - k] = d
-                    I[i - k] = j - k
-                if d < P[j - k]:
-                    P[j - k] = d
-                    I[j - k] = i - k
-
-    return P, I
-
-
-def naive_scraamp(T_A, m, T_B, percentage, exclusion_zone, pre_scraamp, s, p=2.0):
-    distance_matrix = naive.aamp_distance_matrix(T_A, T_B, m, p)
-
-    n_A = T_A.shape[0]
-    n_B = T_B.shape[0]
-    l = n_A - m + 1
-
-    if exclusion_zone is not None:
-        diags = np.random.permutation(range(exclusion_zone + 1, n_A - m + 1)).astype(
-            np.int64
-        )
-    else:
-        diags = np.random.permutation(range(-(n_A - m + 1) + 1, n_B - m + 1)).astype(
-            np.int64
-        )
-
-    n_chunks = int(np.ceil(1.0 / percentage))
-    ndist_counts = core._count_diagonal_ndist(diags, m, n_A, n_B)
-    diags_ranges = core._get_array_ranges(ndist_counts, n_chunks, False)
-    diags_ranges_start = diags_ranges[0, 0]
-    diags_ranges_stop = diags_ranges[0, 1]
-
-    out = np.full((l, 4), np.inf, dtype=object)
-    out[:, 1:] = -1
-    left_P = np.full(l, np.inf, dtype=np.float64)
-    right_P = np.full(l, np.inf, dtype=np.float64)
-
-    for diag_idx in range(diags_ranges_start, diags_ranges_stop):
-        k = diags[diag_idx]
-
-        for i in range(n_A - m + 1):
-            for j in range(n_B - m + 1):
-                if j - i == k:
-                    if distance_matrix[i, j] < out[i, 0]:
-                        out[i, 0] = distance_matrix[i, j]
-                        out[i, 1] = i + k
-
-                    if (
-                        exclusion_zone is not None
-                        and distance_matrix[i, j] < out[i + k, 0]
-                    ):
-                        out[i + k, 0] = distance_matrix[i, j]
-                        out[i + k, 1] = i
-
-                    # left matrix profile and left matrix profile indices
-                    if (
-                        exclusion_zone is not None
-                        and i < i + k
-                        and distance_matrix[i, j] < left_P[i + k]
-                    ):
-                        left_P[i + k] = distance_matrix[i, j]
-                        out[i + k, 2] = i
-
-                    # right matrix profile and right matrix profile indices
-                    if (
-                        exclusion_zone is not None
-                        and i + k > i
-                        and distance_matrix[i, j] < right_P[i]
-                    ):
-                        right_P[i] = distance_matrix[i, j]
-                        out[i, 3] = i + k
-
-    return out
 
 
 test_data = [
@@ -140,7 +33,7 @@ def test_prescraamp_self_join(T_A, T_B):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_P, ref_I = naive_prescraamp(T_B, m, T_B, s=s, exclusion_zone=zone, p=p)
+            ref_P, ref_I = naive.prescraamp(T_B, m, T_B, s=s, exclusion_zone=zone, p=p)
 
             np.random.seed(seed)
             comp_P, comp_I = prescraamp(T_B, m, s=s, p=p)
@@ -158,7 +51,7 @@ def test_prescraamp_A_B_join(T_A, T_B):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_P, ref_I = naive_prescraamp(T_A, m, T_B, s=s, p=p)
+            ref_P, ref_I = naive.prescraamp(T_A, m, T_B, s=s, p=p)
 
             np.random.seed(seed)
             comp_P, comp_I = prescraamp(T_A, m, T_B=T_B, s=s, p=p)
@@ -175,7 +68,7 @@ def test_prescraamp_A_B_join_swap(T_A, T_B):
         seed = np.random.randint(100000)
 
         np.random.seed(seed)
-        ref_P, ref_I = naive_prescraamp(T_B, m, T_A, s=s)
+        ref_P, ref_I = naive.prescraamp(T_B, m, T_A, s=s)
 
         np.random.seed(seed)
         comp_P, comp_I = prescraamp(T_B, m, T_B=T_A, s=s)
@@ -193,7 +86,7 @@ def test_prescraamp_self_join_larger_window(T_A, T_B, m):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_P, ref_I = naive_prescraamp(T_B, m, T_B, s=s, exclusion_zone=zone)
+            ref_P, ref_I = naive.prescraamp(T_B, m, T_B, s=s, exclusion_zone=zone)
 
             np.random.seed(seed)
             comp_P, comp_I = prescraamp(T_B, m, s=s)
@@ -220,7 +113,7 @@ def test_scraamp_self_join(T_A, T_B, percentages):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_mp = naive_scraamp(T_B, m, T_B, percentage, zone, False, None, p=p)
+            ref_mp = naive.scraamp(T_B, m, T_B, percentage, zone, False, None, p=p)
             ref_P = ref_mp[:, 0]
             ref_I = ref_mp[:, 1]
             ref_left_I = ref_mp[:, 2]
@@ -259,7 +152,7 @@ def test_scraamp_A_B_join(T_A, T_B, percentages):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_mp = naive_scraamp(T_A, m, T_B, percentage, None, False, None, p=p)
+            ref_mp = naive.scraamp(T_A, m, T_B, percentage, None, False, None, p=p)
             ref_P = ref_mp[:, 0]
             ref_I = ref_mp[:, 1]
             ref_left_I = ref_mp[:, 2]
@@ -299,7 +192,7 @@ def test_scraamp_A_B_join_swap(T_A, T_B, percentages):
         seed = np.random.randint(100000)
 
         np.random.seed(seed)
-        ref_mp = naive_scraamp(T_B, m, T_A, percentage, None, False, None)
+        ref_mp = naive.scraamp(T_B, m, T_A, percentage, None, False, None)
         ref_P = ref_mp[:, 0]
         # ref_I = ref_mp[:, 1]
         ref_left_I = ref_mp[:, 2]
@@ -335,7 +228,7 @@ def test_scraamp_self_join_larger_window(T_A, T_B, m, percentages):
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_mp = naive_scraamp(T_B, m, T_B, percentage, zone, False, None)
+            ref_mp = naive.scraamp(T_B, m, T_B, percentage, zone, False, None)
             ref_P = ref_mp[:, 0]
             ref_I = ref_mp[:, 1]
             ref_left_I = ref_mp[:, 2]
@@ -507,10 +400,10 @@ def test_scraamp_plus_plus_self_join(T_A, T_B, percentages):
                 seed = np.random.randint(100000)
 
                 np.random.seed(seed)
-                ref_P, ref_I = naive_prescraamp(
+                ref_P, ref_I = naive.prescraamp(
                     T_B, m, T_B, s=s, exclusion_zone=zone, p=p
                 )
-                ref_mp = naive_scraamp(T_B, m, T_B, percentage, zone, True, s, p=p)
+                ref_mp = naive.scraamp(T_B, m, T_B, percentage, zone, True, s, p=p)
                 for i in range(ref_mp.shape[0]):
                     if ref_P[i] < ref_mp[i, 0]:
                         ref_mp[i, 0] = ref_P[i]
@@ -557,8 +450,8 @@ def test_scraamp_plus_plus_A_B_join(T_A, T_B, percentages):
                 seed = np.random.randint(100000)
 
                 np.random.seed(seed)
-                ref_P, ref_I = naive_prescraamp(T_A, m, T_B, s=s, p=p)
-                ref_mp = naive_scraamp(T_A, m, T_B, percentage, None, False, None, p=p)
+                ref_P, ref_I = naive.prescraamp(T_A, m, T_B, s=s, p=p)
+                ref_mp = naive.scraamp(T_A, m, T_B, percentage, None, False, None, p=p)
                 for i in range(ref_mp.shape[0]):
                     if ref_P[i] < ref_mp[i, 0]:
                         ref_mp[i, 0] = ref_P[i]
@@ -691,7 +584,7 @@ def test_scraamp_constant_subsequence_self_join(percentages):
         seed = np.random.randint(100000)
 
         np.random.seed(seed)
-        ref_mp = naive_scraamp(T, m, T, percentage, zone, False, None)
+        ref_mp = naive.scraamp(T, m, T, percentage, zone, False, None)
         ref_P = ref_mp[:, 0]
         # ref_I = ref_mp[:, 1]
         # ref_left_I = ref_mp[:, 2]
@@ -729,7 +622,7 @@ def test_scraamp_identical_subsequence_self_join(percentages):
         seed = np.random.randint(100000)
 
         np.random.seed(seed)
-        ref_mp = naive_scraamp(T, m, T, percentage, zone, False, None)
+        ref_mp = naive.scraamp(T, m, T, percentage, zone, False, None)
         ref_P = ref_mp[:, 0]
         # ref_I = ref_mp[:, 1]
         # ref_left_I = ref_mp[:, 2]
@@ -775,7 +668,7 @@ def test_scraamp_nan_inf_self_join(
             seed = np.random.randint(100000)
 
             np.random.seed(seed)
-            ref_mp = naive_scraamp(T_B_sub, m, T_B_sub, percentage, zone, False, None)
+            ref_mp = naive.scraamp(T_B_sub, m, T_B_sub, percentage, zone, False, None)
             ref_P = ref_mp[:, 0]
             ref_I = ref_mp[:, 1]
             ref_left_I = ref_mp[:, 2]
@@ -809,7 +702,7 @@ def test_scraamp_nan_zero_mean_self_join(percentages):
         seed = np.random.randint(100000)
 
         np.random.seed(seed)
-        ref_mp = naive_scraamp(T, m, T, percentage, zone, False, None)
+        ref_mp = naive.scraamp(T, m, T, percentage, zone, False, None)
         ref_P = ref_mp[:, 0]
         ref_I = ref_mp[:, 1]
         ref_left_I = ref_mp[:, 2]
